tests: Integrate existing math tests with test framework
Math tests can now be run with 'make check'

Create one binary per test.
Embed test data with binary.
Use runtest to start tests.

The  three above should make it easier to add support for running
the tests on an accelerator (e.g. Epiphany).

Note: In the time in between I started implementing the code and
writing this commit log the API specification did undergo a major
change.
[[ a380f6b70b8461dbb8c0def388d00270f8b27c28 ]]
But the implementation is still the same as the old API. So we rely on
a hack for now (pass in an invalid team) which means tests will
to break when the new API is implemented.

Signed-off-by: Ola Jeppsson <ola@adapteva.com>