Add p_sort() to math library (with an example)
This commit adds p_sort_f32 and p_sort_u32 to the math library,
as well as example (basic test code).

The prototype of p_sort_u32 was changed from `int` to `uint32_t`
for the sake of consistency.

The implementation via two larg-ish macros was chosen to keep the
code (which is mostly the same across both functions) DRY.

Issue referenced: #48.

Signed-off-by: Michal Jirku <box@wejn.org>