Image: filters example (read/write image)
This example reads an image (lena.tga), converts it to float/greyscale
and runs the following filters :

- gaussian > saved to lena_gaussian.tga
- harris > saved to lena_harris.tga
- sobel > saved to lena_sobel.tga (broken > see image)
- box > saved to lena_box.tga
- median > saved to lena_median.tga (wrong result I think)

I pushed the filtered images so you can see the results easily, but
maybe it should be arranged in another folder.

I didn’t write stb_image, it is a public domain library.

Signed-off-by: Anaël Seghezzi anael@maratis3d.com