Branchless version of function
Get the sign bit from the input float and set that sign bit on 0.5f rounding value so we can always add the rounding value to the input instead of checking the sign and branching to add or subtract 0.5f.  Used #define for the magic number to help be more self documenting.  Don't know if that fits the style or not, easy to take it out.