dsp: p_fir: Fix FIR filter implementation
The new implementation follows the same idea as the previous one: simultaneously processes blocks of 8 samples to leverage vectorization. On an i7-3770 PC, this function is almost 4 times faster than the evaluating the output sample-by-sample (compiled with -O3). Also removed the reference to dbuf from the function description as it does not exist anymore (it was removed from the API in commit #1fa28ced55ebddb522cd383ce41bf0a792aff177).

Signed-off-by: Luiz Chamon <luizchamon@gmail.com>