p_mutex_*: Implement init/lock/unlock/etc:
Implement the p_mutex functions.

p_mutex_t is a uint32_t* -- We waste 23 bytes on it, but it's a
nice round size.

This follows in the form that pthreads uses. Each of these functions
does a quick NULL check to be sure things aren't bad (except for
p_mutex_create, which checks that we're not trying to initialize an
empty p_mutex_t)

Signed-off-by: Morgan Gangwere <morgan.gangwere@gmail.com>