p_memcpy: fix invalid assertion logic
Bad assertion logic would have caused a problem if `dst < src`
in debug builds. This logic would have also caused a problem
if the flag FAIL_FAST was defined at compile-time.

This should probably be wrapped in a macro (or some such) that
checks that the bounds are correct. Multiple asserts in the non
fail-fast variant are cleaner, but may or may not be optimized
out when gcc comes along and tries to run "basic" optimizations.

Signed-off-by: Morgan Gangwere <morgan.gangwere@gmail.com>