Implement fastest pow() function base on the work of Harrison Ainsworth
With a precision table set to 11 bits (8KB table), mean error is < 0.01%, and max error is < 0.02%.
On my computer i7-2630QM @ 2.0GHz I observe the code running 11 times faster than the glibc pow().

This implementation is based on work of Harrison Ainsworth.
http://www.hxa.name/articles/content/fast-pow-adjustable_hxa7241_2007.html

Golden have been modified as the new pow() is an approximation function.

Signed-off-by: Nicolas PLANEL <nplanel@redhat.com>