tests: Add uTest framework
Nothing bad about 'check' but it is not well suited for embedded
and I didn't want to go trough the endeavour to port it to Epiphany.

The code size clocks in at circa 39kB (add another 20 kB if you want
to call printf()) on Epiphany which is still too big to fit in core
memory. Most of that comes from all the stuff snprintf()/vsnprintf()
drags in. Without it the code size is about 1132 for the uTest library
functions itself, which is not the fairest way to count but gives a
good indication over how light it is.

The library distinguishes between execute() and verify() which will
make easier to incorporate benchmarking in the tests, which I think is
the right way to go.

Signed-off-by: Ola Jeppsson <ola@adapteva.com>