tests:utest: [2/2] Move ut_report to own file
Instead of rewriting ut_report() to not use snprintf() and replacing it
with something else, move it to a separate compilation unit so that the
linker does not have to bring in its dependencies for programs that
never call it. We're now down to 3100 bytes for a complete binary on
Epiphany.

Signed-off-by: Ola Jeppsson <ola@adapteva.com>