image: p_sobel3x3: new algrthm achieves 76% peak
The core of the routine achieves 9+ GOPS (over 330 MPixels/s) and
this is ~76% of theoretical performance using a roofline model based
on the multiply/add instruction imbalance. The algorithm improves
upon the earlier code by storing accumulated 3-value columns in
registers as it processes each row. This is similar to the Prewitt
and Scharr operators.