image: p_prewitt3x3: new algrthm achieves 71% peak
The core of the routine achieves 7.7+ GOPS (over 308 MPixels/s) and
this is ~71% of theoretical performance using a roofline model based
on the multiply/add instruction imbalance. The algorithm improves
upon the earlier code by storing accumulated 3-value columns in
registers as it processes each row. This is similar to the Sobel
and Scharr operators.