image: p_scharr3x3: new algrthm 76% peak + bugfix
There was a bug in original code. Constant should be 1/32, not 1/36.
The core of the routine achieves 9+ GOPS (over 330 MPixels/s) and
this is ~76% of theoretical performance using a roofline model based
on the multiply/add instruction imbalance. The algorithm improves
upon the earlier code by storing accumulated 3-value columns in
registers as it processes each row. This is similar to the Sobel
and Prewitt operators.