image: p_gauss3x3: new algrthm 63% peak + consts
The Gaussian constants were incorrect and a more details may
be found in the code comments.
The core of the routine achieves 8+ GOPS (over 1 GigaPixel/s) and
this is ~63% of theoretical performance using a roofline model based
on the multiply/add instruction imbalance. The algorithm improves
upon the earlier code by storing accumulated 3-value columns in
registers as it processes each row.