math:p_[inv]cbrt:impl fast approx [inv]cbrt
The inverse cube root and cube root routines are similar as the inverse is
easier to calculate and the cube root can be derived from it: x^(1/3) =
x*x^(-1/3)*x^(-1/3)

Emperical results will likely show that the magic numbers are not optimal, but
they are close and the initial approximation is not as critical as the Newton
iterations for accuracy.

Both routines appear robust against a range of input and correctly handle
negative input values.  No division operations are required so they should be
faster for the Epiphany than other algorithms which require division.

Signed-off-by: James Ross james.a.ross@gmail.com