devices: epiphany: Look up addresses by section where applicable
Look up addresses in ELF file by section instead of by symbol.
The linker scripts will always output the section even when the symbol
is not present (beacuse the binary doesn't use e-lib).

This is not ideal, we should support all binaries.
E.g., programs not linked against -lpal will not signal the host when it
terminates. We need a fallback for that (check for
__pal_init/__pal_fini and check debugstatus + insn at PC in epiphany
p_wait).

Signed-off-by: Ola Jeppsson <ola@adapteva.com>