devices/epiphany:sim: Delay connect call to simulator to init
Move the call to es_client_connect() from early_init() to init().
early_init() is called from a constructor (init section). I.e,
for programs linked against PAL, es_client_connect() is called
before main() is entered.

This interferes with e-hal, which might also want to connect
to the simulator and not use the PAL backend.

We should be restrictive about what we put in early_init.

Signed-off-by: Ola Jeppsson <ola@adapteva.com>