epiphany: Use team API for device code
Adds support for 2D team topology on for Epiphany as device.
Let the host loader fill in the __pal_epiphany_coords structure.
Add packed attribute to the p_coords_t type so we reliably can pass data
between architectures. Since struct team / struct device are not packed
we cannot easily poke the values in directly.
As a side effect we also finally get rid of the e-lib dependency in this
commit.

Signed-off-by: Ola Jeppsson <ola@adapteva.com>